fix(browser): keep the session profile selection on the client that made it
Review follow-ups on #19930.

The persisted selection crossed the pairing boundary. Profile ids are minted
per machine and their cookie jars live there, so a paired client's choice
overwrote every other client's, leaving their Settings pane with no Active
row. Classify the field as pairing-local and pin it on the web read path.

Deleting a profile decided whether to persist the cleared selection from a
snapshot taken before the awaited delete request. Selecting that profile
while the request was in flight cleared the store without persisting it, so
the next launch restored a profile that no longer exists. Read the selection
after the request instead.

The e2e helper relied on timing: the store persists through a fire-and-forget
ui.set, so the relaunch assertion could race the IPC it depends on. Wait for
the choice to reach persisted state before quitting.

Split the web preload UI spec: it was already at the max-lines limit before
this branch, and the pairing-local tests are a self-contained contract.
